introduced solid tumor cancer base class and supports grade as a number now. also added race retrieval from patient
